+++ extensions/dust-tt/src/dust_api/api.tsx
@@ -7,7 +7,7 @@ import {
   AgentActionSuccessEvent,
   AgentActionType,
   AgentErrorEvent,
-  AgentGenerationSuccessEvent,
+  AgentMessageSuccessEvent,
   DustAPIErrorResponse,
   DustDocument,
   GenerationTokensEvent,
@@ -95,7 +95,7 @@ export class DustApi {
 
     if (!res.ok || !res.body) {
       console.error(`Error running streamed app: status_code=${res.status}  - message=${await res.text()}`);
-      throw new Error(`Error running streamed app: status_code=${res.status}`);
+      return null;
     }
 
     const pendingEvents: (
@@ -103,7 +103,7 @@ export class DustApi {
       | AgentErrorEvent
       | AgentActionSuccessEvent
       | GenerationTokensEvent
-      | AgentGenerationSuccessEvent
+      | AgentMessageSuccessEvent
     )[] = [];
     const parser = createParser((event) => {
       if (event.type === "event") {
@@ -137,7 +137,6 @@ export class DustApi {
       });
       reader.on("error", (err) => {
         console.error("Error reading stream", err);
-        throw err;
       });
       while (!done) {
         if (pendingEvents.length > 0) {
@@ -177,15 +176,12 @@ export class DustApi {
         });
       if (agentMessages.length === 0) {
         console.error("Failed to retrieve agent message");
-        throw new Error("Failed to retrieve agent message");
       }
       const agentMessage = agentMessages[0];
-      const streamRes = agentMessage
-        ? await this.streamAgentMessageEvents({
-            conversationId,
-            messageId: agentMessage.sId,
-          })
-        : undefined;
+      const streamRes = await this.streamAgentMessageEvents({
+        conversationId,
+        messageId: agentMessage.sId,
+      });
       if (!streamRes) {
         return;
       }
@@ -199,17 +195,20 @@ export class DustApi {
         switch (event.type) {
           case "user_message_error": {
             console.error(`User message error: code: ${event.error.code} message: ${event.error.message}`);
-            throw new Error(event.error.message, { cause: event.error.code });
+            return;
           }
           case "agent_error": {
             console.error(`Agent message error: code: ${event.error.code} message: ${event.error.message}`);
-            throw new Error(event.error.message, { cause: event.error.code });
+            return;
           }
           case "agent_action_success": {
             action = event.action;
             break;
           }
           case "generation_tokens": {
+            if (event.classification !== "tokens") {
+              continue;
+            }
             answer += event.text;
             if (lastSentDate.getTime() + 500 > new Date().getTime()) {
               continue;
@@ -219,8 +218,8 @@ export class DustApi {
             setDustAnswer(dustAnswer + "...");
             break;
           }
-          case "agent_generation_success": {
-            answer = this.processAction({ content: event.text, action, setDustDocuments });
+          case "agent_message_success": {
+            answer = this.processAction({ content: event.message.content ?? "", action, setDustDocuments });
             setDustAnswer(answer);
             if (onDone) {
               onDone(answer);
@@ -298,17 +297,12 @@ export class DustApi {
     const { apiKey, workspaceId } = this._credentials;
     const agentsUrl = `${DUST_API_URL}/${workspaceId}/assistant/agent_configurations`;
 
-    let response;
-    try {
-      response = await fetch(agentsUrl, {
-        method: "GET",
-        headers: {
-          Authorization: `Bearer ${apiKey}`,
-        },
-      });
-    } catch (error) {
-      return { error: `Could not get agents: ${error}` };
-    }
+    const response = await fetch(agentsUrl, {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${apiKey}`,
+      },
+    });
     if (!response.ok) {
       return { error: `Could not get agents: ${response.statusText}` };
     }