+++ extensions/github/src/my-issues.tsx
@@ -1,9 +1,11 @@
 import { List } from "@raycast/api";
+import { useCachedState } from "@raycast/utils";
 import { useState } from "react";
 
 import IssueListEmptyView from "./components/IssueListEmptyView";
 import IssueListItem from "./components/IssueListItem";
 import RepositoriesDropdown from "./components/RepositoryDropdown";
+import { ISSUE_DEFAULT_SORT_QUERY } from "./helpers/issue";
 import { withGitHubClient } from "./helpers/withGithubClient";
 import { useMyIssues } from "./hooks/useMyIssues";
 import { useViewer } from "./hooks/useViewer";
@@ -11,7 +13,10 @@ import { useViewer } from "./hooks/useViewer";
 function MyIssues() {
   const viewer = useViewer();
   const [selectedRepository, setSelectedRepository] = useState<string | null>(null);
-  const { data: sections, isLoading, mutate: mutateList } = useMyIssues(selectedRepository);
+  const [sortQuery, setSortQuery] = useCachedState<string>("sort-query", ISSUE_DEFAULT_SORT_QUERY, {
+    cacheNamespace: "github-my-issue",
+  });
+  const { data: sections, isLoading, mutate: mutateList } = useMyIssues(selectedRepository, sortQuery);
 
   return (
     <List
@@ -23,7 +28,7 @@ function MyIssues() {
         return (
           <List.Section key={section.title} title={section.title} subtitle={section.subtitle}>
             {section.issues?.map((issue) => {
-              return <IssueListItem key={issue.id} issue={issue} viewer={viewer} mutateList={mutateList} />;
+              return <IssueListItem key={issue.id} {...{ issue, viewer, mutateList, sortQuery, setSortQuery }} />;
             })}
           </List.Section>
         );