+++ extensions/github/src/search-issues.tsx
@@ -1,5 +1,5 @@
 import { getPreferenceValues, List } from "@raycast/api";
-import { useCachedPromise } from "@raycast/utils";
+import { useCachedPromise, useCachedState } from "@raycast/utils";
 import { trim } from "lodash";
 import { useState } from "react";
 
@@ -10,6 +10,7 @@ import { getBoundedPreferenceNumber } from "./components/Menu";
 import SearchRepositoryDropdown from "./components/SearchRepositoryDropdown";
 import { IssueFieldsFragment } from "./generated/graphql";
 import { pluralize } from "./helpers";
+import { ISSUE_DEFAULT_SORT_QUERY } from "./helpers/issue";
 import { withGitHubClient } from "./helpers/withGithubClient";
 import { useViewer } from "./hooks/useViewer";
 
@@ -20,6 +21,9 @@ function SearchIssues() {
 
   const { defaultSearchTerms } = getPreferenceValues<Preferences>();
   const [searchText, setSearchText] = useState(trim(defaultSearchTerms) + " ");
+  const [sortQuery, setSortQuery] = useCachedState<string>("sort-query", ISSUE_DEFAULT_SORT_QUERY, {
+    cacheNamespace: "github-search-issue",
+  });
   const [searchFilter, setSearchFilter] = useState<string | null>(null);
 
   const {
@@ -27,15 +31,15 @@ function SearchIssues() {
     isLoading,
     mutate: mutateList,
   } = useCachedPromise(
-    async (searchText, searchFilter) => {
+    async (searchText, searchFilter, sortTxt) => {
       const result = await github.searchIssues({
         numberOfItems: getBoundedPreferenceNumber({ name: "numberOfResults", default: 50 }),
-        query: `is:issue archived:false ${searchFilter} ${searchText}`,
+        query: `is:issue archived:false ${sortTxt} ${searchFilter} ${searchText}`,
       });
 
       return result.search.nodes?.map((node) => node as IssueFieldsFragment);
     },
-    [searchText, searchFilter],
+    [searchText, searchFilter, sortQuery],
     { keepPreviousData: true },
   );
 
@@ -54,7 +58,7 @@ function SearchIssues() {
           subtitle={pluralize(data.length, "issue", { withNumber: true })}
         >
           {data.map((issue) => {
-            return <IssueListItem key={issue.id} issue={issue} viewer={viewer} mutateList={mutateList} />;
+            return <IssueListItem key={issue.id} {...{ issue, viewer, mutateList, sortQuery, setSortQuery }} />;
           })}
         </List.Section>
       ) : null}