+++ extensions/github/src/search-repositories.tsx
@@ -1,5 +1,5 @@
 import { List, getPreferenceValues } from "@raycast/api";
-import { useCachedPromise } from "@raycast/utils";
+import { useCachedPromise, useCachedState } from "@raycast/utils";
 import { useEffect, useMemo, useState } from "react";
 
 import { getGitHubClient } from "./api/githubClient";
@@ -8,7 +8,7 @@ import RepositoryListEmptyView from "./components/RepositoryListEmptyView";
 import RepositoryListItem from "./components/RepositoryListItem";
 import SearchRepositoryDropdown from "./components/SearchRepositoryDropdown";
 import { ExtendedRepositoryFieldsFragment } from "./generated/graphql";
-import { useHistory } from "./helpers/repository";
+import { REPO_DEFAULT_SORT_QUERY, REPO_SORT_TYPES_TO_QUERIES, useHistory } from "./helpers/repository";
 import { withGitHubClient } from "./helpers/withGithubClient";
 
 function SearchRepositories() {
@@ -18,14 +18,18 @@ function SearchRepositories() {
 
   const [searchText, setSearchText] = useState("");
   const [searchFilter, setSearchFilter] = useState<string | null>(null);
+  const [sortQuery, setSortQuery] = useCachedState<string>("sort-query", REPO_DEFAULT_SORT_QUERY, {
+    cacheNamespace: "github-search-repo",
+  });
+  const sortTypesData = REPO_SORT_TYPES_TO_QUERIES;
 
   const { data: history, visitRepository } = useHistory(searchText, searchFilter);
   const query = useMemo(
     () =>
-      `${searchFilter} ${searchText} sort:updated-desc fork:${preferences.includeForks} ${
+      `${searchFilter} ${searchText} ${sortQuery} fork:${preferences.includeForks} ${
         preferences.includeArchived ? "" : "archived:false"
       }`,
-    [searchText, searchFilter],
+    [searchText, searchFilter, sortQuery],
   );
 
   const {
@@ -67,9 +71,7 @@ function SearchRepositories() {
         {history.map((repository) => (
           <RepositoryListItem
             key={repository.id}
-            repository={repository}
-            onVisit={visitRepository}
-            mutateList={mutateList}
+            {...{ repository, onVisit: visitRepository, mutateList, sortQuery, setSortQuery, sortTypesData }}
           />
         ))}
       </List.Section>
@@ -83,9 +85,7 @@ function SearchRepositories() {
             return (
               <RepositoryListItem
                 key={repository.id}
-                repository={repository}
-                mutateList={mutateList}
-                onVisit={visitRepository}
+                {...{ repository, onVisit: visitRepository, mutateList, sortQuery, setSortQuery, sortTypesData }}
               />
             );
           })}