+++ extensions/devdocs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # DevDocs Extension
 
+## [Changes] - 2024-07-21
+
+- Update icon and include dark mode version
+- Update command name and copy to match "documentations" terminology for consistency with DevDocs
+- Bump Raycast API and utils
+- Add shortcut to create Quicklink
+- Add action to copy documentation slug to clipboard
+- Clarify default Quicklink name
+- Add navigation reference to current documentation
+
 ## [Fix] - 2023-03-14
 
 - Fixed wrong link in `Open in DevDocs`